reorganise tests folder
